updated deployment action
used a version copied from another project (URL in source). 

one key change is the use of the `JamesIves/github-pages-deploy-action@3.7.1` action. in my previous attemps, this path looked like the name of the current repository, so i changed it to "connectuum/poemdown@main". now, however, it looks to me like this action is required -- it likely contains the steps/logic for deploying to GitHub Pages.