flowed input into html conversion into overlay view
finally, we're into some more interesting work. glad to be through the frustration of deployment actions. the need to commit each test-trial into the repository, combined with the long turnaround time between each test, combined with the guess-the-magic-numbers feeling of configuration files, made it an especially annoying task.

so: _thrilled_ to be through with initial project setup and config. getting started with work on an idea is difficult enough on its own. and for the moment, it seems the headwinds will relent.

within this commit, i decided on a first task: read and react to input text. 

for the reading: an event handler collects the current text upon each change. this stumped me for a bit. i initially used the `onchange` handler, but received no events. i side-tracked into reviewing/debugging the handler function itself, but eventually returned to the source of the problem: it demands the `oninput` handler.

for the reaction: added a static class for converting poemdown text into other formats. for now, we just need `ToHtml()`. i had a little fun with it. and then _Hourglass_ came on while i was working here, and some comments spilled out.

seeing the lyrics as "two" jumped at me. [mitosis monkey, right in two]. human systems express stress as sharpening spears of opposition and fear. one way to inoculate: become more deeply interconnected. connection provides the strain of context. it resists spiky behaviors. [you pull the elastic], [the elastic pulls back].

after the conversion, the resulting HTML flows into the `div.overlay`, which renders exactly on top of the `textarea`. i'm not sure if this approach will be sufficient, but i'd rather not bring in a third-party library for this. i think we'll just need syntax highlighting here (and not all the text-formatting buttons that most of these libraries would provide).

bug: newline-to-break conversion
soon: regular expressions
horizon: d3 and force-directed graphs