added support for tab-character input
the tab key's default behavior, in a web browser, is to shift focus between input fields. for now, this app only has one input field, and when the user is typing, we need the tab key to create whitespace. to handle this, we capture the `keydown` event, prevent its default behavior, and update the input text to include a tab (`\t`) character.

unfortunately, the `textarea` does not automatically receive an input event when the tab character is added. i figured out how to get the C# `onkeydown` event handler to work. now, any change will trigger a new HTML conversion (assuming the input text has changed since the last conversion).

we also need special handling of tabs in the conversion process. the converted text must consider the tab _stops_, not just the tab _size_.

lastly, the `*.lib.module.js` is new. this is a specially-named JS file (with specially-named functions) that Blazor uses to notify our JS code when the Blazor-created HTML is rendered and ready for us.