improved the loading view
the top-left-corner "Loading..." label had to go. the loading takes a while, and i couldn't bear to stare at it anymore. as part of this change, the "fullscreen" div moved into the primary HTML structure of the app.

todo: see what can be done about that load time. i've read that Blazor can do some lazy-loading during startup. and this app may not need most of the things Blazor loads by default, anyway.