wrote initial "markdown" rules; added console app
today, brave souls, we battle a most devious foe: _regular expressions._ if you can withstand their powers of confusion, great is the reward.

i am pleased to tell you that i do _not_ need to teach you how to use them. i'll happily tell you _about_ them, though.

`PoemdownConvert.cs:19`: here begins the code block (until line 26) where you'll find the fantastic tangle of symbols that are named "regular expressions". the `Replace()` function takes (text-to-search, pattern-to-match, replacement-text-for-matches) input parameters, and returns (text-with-all-matches-replaced). regular expressions represent a _pattern_ or _key_, used for finding matches within any given text.

if you wish to take a peek behind their curtain, search for a regular expression _interpreter_. copy all the symbols within the `"` quotes. for example: the pattern part of `@"(_[^_\n]+_)"`  starts and ends at the `(` and `)` parentheses. paste that pattern into the interpreter and behold the absolute wonders of this tiny piece of text.

for those that dare to learn, regular expressions are an incredibly compact, incredibly powerful tool to wield upon text. in the space of a few delicately-selected keystrokes, you can immediately scan _any_ passage of text for matches to your pattern, snip _out_ the text of each match, and splice _in_ a per-match-customizable replacement texts.

and of course, _text_ is not limited to _sentences, etc._ for regular expressions and computers, _text_ (a.k.a. `string`) does not have any meaning beyond a simple a list of characters. each character, one little event after another, arranged forever in this order, awaiting the RegExp ReadHead.

and what is a character to a computer? it is simply a number. and what is that number? it is simply a list of bits. the `0` and the `1`. a lovely image of the duality of all things. the computer knows them only as _off_ and _on_, _resist_ or _flow_. _although:_ it knows those things at untold billions of times per second, and will execute our every instruction with an unimaginable capacity for precision.

_tangent:_ when you get angry at the supercomputer in your pocket for freezing up during your nineteenth game of _whatever_, you probably curse the machine. don't be foolish. the machine is doing _exactly_ what it has been told to do. by us. it's just that we're not always very good at the _telling_. or, we _are_ pretty good at it, _but_, your pocket supercomputer is executing so very many things all the time, and conflicts arise. one particularly pernicious conflict that can cause a freeze is called a _deadlock_: at the same time, two different tasks both request exclusive access to some data, and upon hearing of the other, each politely await their turn until the end of time (or until they are `kill`ed by their fleshy, unfathomably biological god).

_recursion up:_ and probably up again, there must have been another tangent in there at some point. so, _regular expressions:_ a function which takes (your-careful-keystrokes) as an input parameter and returns (a-million-years-of-search-and-replace-by-hand-with-zero-mistakes-in-fraction-of-a-second).

computers are fantastically powerful _tools._ use them. for _good._

**stop letting them use _you._**