improved the overlay HTML and styles
the overlay colors are now more vibrant, while the markdown symbols are more subdued. the goal is for the control characters to be present, readable, editable, but _not_ outshining or distracting from the words. this is the editor panel, so we need to see everything clearly -- but not necessarily at equal weighting.

also, we now use custom HTML tags (like `head1`) in place of `span` tags with  `class` specifications. not a vital change, but it looks cleaner in the code, and it avoids the need for the `\v` special case.