added new Poemdown.Core project
a few reasons for this change:

the `Poemdown.Blaze` project is primarily meant to _display_ content as (and accept user input from) a website. it does not need to host any of the general poemdown logic and routines.

a class library project is easier to use (and is more efficient) than a Blaze-based project. when testing via the `Poemdown.Console` project, we will now only need to wait for the `Poemdown.Core` project to build, which is a faster operation than building a full Blaze-based project.

a class library is naturally reusable. if it contains all the "core" logic and routines that poemdown requires, then we could reuse this library for any new display-related projects we might want to create. for example, we might want to build a display for a Unity project, or even a terminal-based app. all of these displays can depend upon the `.Core` project, but have no dependencies between them (i.e. a `.Terminal` project would have no knowledge of, or dependencies upon, the `.Unity` or `.Blaze` projects).