fixed word-wrap issue (#3)
this fix has a few different parts.

most importantly, the use of `&nbsp;` in place of simple ` ` space characters was disrupting the natural word-wrapping. the HTML conversion now avoids this replacement in single-space situations. the `&nbsp;` symbol forces space character(s) to appear in HTML; typically HTML collapses multiple spaces into a single one.

by adding an `overflow-wrap` style to the overlay, we also handle scenarios where the text is so long that it must forcibly break.

applying the `display: block` style to `textarea` isn't related to the fix, but it _does_ naturally improve alignment, allowing us to remove the `margin-top: -2px` "nudge" from the overlay style.